Fix 2 issues, including the problem of excessively large directories.
1) Now stores files in a tiered set of subdirectories of the form
"a/b/owner", where "a" is the first character of the owner's name and
"b" is the second character.  This works around an inconvenience in
Unix "ls" where a naive "ls" on a very large directory will take
forever to return a result.

2) Trying to work around codec problems on Linux.  The problem has
been that sometimes you get the following error:

  4Lang/4: /srv/repositories/4Lang/4/4-master.zip left unzipped:
  'ascii' codec can't encode character '\U0001f340' in position 43:
  ordinal not in range(128)

The problem is coming from the zipfile library when it tries to unzip
the file it receives from GitHub.  The problem does not exist if you
run the command-line program "unzip", which means the culprit is the
Python zipfile library.

Some Googling around suggests the problem may be related to setting
the system locale, so that's what I tried to do here.