Add separate command-line interface script
This is part of a reorganization of the code, following my realization
that I've been doing things kind of un-Pythonically and
inconsistently.  The new approach involves putting command-line
interface scripts in a /bin directory and making that use Python code
in the /modulename directory.  It involves some changes to the module
main file too.